[Gepardec/mega#674] move bills impl to worker resource
